--- a/WebDriverAgentLib/Categories/XCUIElement+FBUtilities.m
+++ b/WebDriverAgentLib/Categories/XCUIElement+FBUtilities.m
@@ -74,11 +74,9 @@ static const NSTimeInterval FBANIMATION_TIMEOUT = 5.0;
 {
   if ([self isKindOfClass:XCUIApplication.class]) {
     [((XCUIApplication *)self) fb_resolve];
-  }
-  else if (self.application.fb_isActivateSupported) {
+  } else if (self.application.fb_isActivateSupported) {
     [self.application fb_resolve];
-  }
-  else {
+  } else {
     [self query];
     [self resolve];
   }