Idea taken from this stackoverflow post https://stackoverflow.com/questions/35295090/possible-to-bring-the-app-from-background-to-foreground. I have made changes to call activate instead of query + resolve for Getting the Page source in iOS11, there is a performance increase

This is my first go at coding in objective-c so be kind :worried: