@tmulkern 2018-01-09
This PR might help in this issue https://github.com/facebook/WebDriverAgent/issues/639