@mykola-mokhnach 2018-01-09
I assume we the performance problem there, since it takes time to read all the attributes (especially `visible` one). Have you tested this PR with any large application? What is the actual perf diff there in comparison to the current implementation?