@tmulkern 2018-01-09
Yes, I have tested on large views (lots of elements) but I don't have any hard numbers. I'll take some measurements tomorrow.