@tmulkern 2018-01-09
@mykola-mokhnach I have further PR's to create with other performance improvements, hopefully will get them in tomorrow too