@tmulkern 2018-01-11
@mykola-mokhnach I have some timings for this:

**Before Change**
XMLSource - 75 seconds average
JSONSource - 75 seconds average

**After Change**
XMLSource - 19 seconds
JSONSource - 19 seconds 

The first call to get source was as before (75 seconds), but every subsequent call was much faster (6-8 seconds)