@mykola-mokhnach 2018-01-11
I've modified the original method like this

```objc
- (XCElementSnapshot *)fb_lastSnapshot
{
  if ([self isKindOfClass:XCUIApplication.class] && ((XCUIApplication *)self).fb_isActivateSupported) {
    [((XCUIApplication *)self) fb_activate];
  } else {
    [self query];
    [self resolve];
  }
  return self.lastSnapshot;
}
```

but I still don't see any visible speed difference in getting the source, at least with my app