@tmulkern 2018-01-11
@mykola-mokhnach Just to explain how I got the timings, I called getPageSource five times in a row.
Without this change, it averages 75 seconds, with the change the first call is approx 75 seconds then every subsequent call is between 6-8 seconds.

Here is an example run:
Time 1: 74649 ms
Time 2: 7977 ms
Time: 3 7064 ms
Time 4: 6056 ms
Time 5: 6197 ms
Avg Time: 20388.6 ms