@tmulkern 2018-01-11
@mykola-mokhnach Also the code changed won't fix the issue in  [FBApplication.fb_activeApplication](https://github.com/facebook/WebDriverAgent/pull/838/files#diff-737b59e6c9285f94225421e7c9404779L44), which calls 

```
[self query];
[self resolve];
```

explicitly