@tmulkern 2018-01-17
@mykola-mokhnach I took on board your changes at re-wrote fb_resolve but put the logic into another method

With these additional change the performance is better again. 

Here are the latest timings: 
Time 1: 73235
Time 2: 3102
Time 3 3033
Time 4: 2695
Time 5: 3391
Avg Time: 17091.2