@tmulkern 2018-01-19
@mykola-mokhnach Ok the results are in...

Time 1: 57572 ms
Time 2: 58447 ms
Time 3: 57679 ms
Time 4: 57344 ms
Time 5: 57548 ms
Avg Time: 57718 ms

I am going to see it I can take the changes from this PR and merge them with the changes you added in the Appium Fork and run the timings again