@mykola-mokhnach 2018-01-19
I actually did too many changes there :) But thanks for trying it out