@tmulkern 2018-01-19
@mykola-mokhnach It knocked 20 secs of the time, not too shabby :wink: