@tmulkern 2018-01-19
@mykola-mokhnach 
I applied the changes from this [commit](https://github.com/appium/WebDriverAgent/commit/02887c171699821314c1928642303c36f037ae8a)
and from this [commit] (https://github.com/appium/WebDriverAgent/commit/02887c171699821314c1928642303c36f037ae8a)
To the changes in this PR and here are the timings

Time 1: 67246 ms
Time 2: 811 ms
Time 3: 571 ms
Time 4: 746 ms
Time 5: 410 ms 
Avg Time: 13956.8 ms