@tmulkern 2018-01-19
@mykola-mokhnach So is this PR OK to merge as is??