@tutipeti 2018-01-31
Any updates on this one? @mykola-mokhnach can you maybe poke whoever actually works for Facebook and can click the merge button? Thanks! :)