@marekcirkos 2018-02-28
Sorry I completely missed this. Can you rebase this diff?

Meanwhile, I have a question. Why you force activate application if it is not active?
Why it helps in performance?