@kimbo-slice 2018-03-31
Howdy y'all! I wanted to see if there was any movement on this PR. My team would be super happy to see our iPad tests see these speed improvements :)