@tmulkern 2018-03-31
@marekcirkos Sorry, I have been away from this work, will rebase this next week.