@mykola-mokhnach 2018-05-11
https://github.com/appium/WebDriverAgent/pull/80 fixes the slowness