@tmulkern 2018-05-18
@mykola-mokhnach I am happy to close as [appium#80](https://github.com/appium/WebDriverAgent/pull/80) solves the route cause. The fix I had doesn't take into account page changes ( I found out later :confounded: ) the call to `[self query];
    [self resolve]` guarantees the snapshot is up to date.